% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboVis.R
\name{metaboVis}
\alias{metaboVis}
\title{metaboVis}
\usage{
metaboVis(tool, ...)
}
\arguments{
\item{tool}{A string denoting which visualisation tool to use}

\item{...}{Parameters to be passed on to selected tool}
}
\description{
metaboVis
}
\examples{
\dontrun{
metaboVis('PCA')
}
}
