% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseAnalysis.R
\name{visualiseAnalysis}
\alias{visualiseAnalysis}
\title{A shiny application for the visualisation of HRfiemsWorkflow analyses}
\usage{
visualiseAnalysis(analysis)
}
\arguments{
\item{analysis}{analysis list object.}
}
\description{
A shiny application for the visualisation of HRfiemsWorkflow analyses
}
\details{
A tool that allows the visualisation of analysis results. Includes tools for TIC visualisation, PCA, LDA, classification and feature selection results visualisation and tools for putative annotation.
Annotation tools use those in the mzAnnotation package (\url{https://github.com/jasenfinch/mzAnnotation}).
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
